refactor: move parser tests from parsers.rs to tests.rs
Consolidates all SysML AST tests in one file for better organization.
